DA Protocol uses Runtime for Extension
