+++ node/Cargo.toml
@@ -18,7 +18,6 @@ name = "data-avail"
 
 [dependencies]
 avail-base = { path = "../base" }
-kate-rpc-runtime-api = { path = "../rpc/kate-rpc-runtime-api" }
 
 # Internals
 da-primitives = { version = "0.4", default-features = false }
