+++ node/src/command.rs
@@ -63,7 +63,7 @@ impl SubstrateCli for Cli {
 	}
 
 	fn native_runtime_version(_: &Box<dyn ChainSpec>) -> &'static RuntimeVersion {
-		&da_runtime::VERSION
+		da_runtime::apis::NATIVE_VERSION
 	}
 }
