+++ rpc/kate-rpc/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 
 [dependencies]
 avail-base = { path = "../../base", default-features = false }
-kate-rpc-runtime-api = { path = "../kate-rpc-runtime-api", default-features = false }
 da-runtime = { path = "../../runtime", default-features = false }
 
 da-primitives = { version = "0.4", default-features = false }
@@ -32,7 +31,6 @@ default = ["std"]
 std = [
 	"avail-base/std",
 	"da-runtime/std",
-	"kate-rpc-runtime-api/std",
 	"da-primitives/std",
 	"kate/std",
 	"kate-recovery/std",
