+++ runtime/Cargo.toml
@@ -18,7 +18,6 @@ targets = ["x86_64-unknown-linux-gnu"]
 da-primitives = { version = "0.4", default-features = false }
 kate = { version = "0.7" , default-features = false }
 da-control = { path = "../pallets/dactr", default-features = false }
-kate-rpc-runtime-api = { path = "../rpc/kate-rpc-runtime-api", default-features = false }
 
 ## Nomad
 nomad-updater-manager = { path = "../pallets/bridges/nomad/updater-manager", default-features = false }
@@ -183,7 +182,6 @@ std = [
 
 	"da-control/std",
 	"kate/std",
-	"kate-rpc-runtime-api/std",
 	"nomad-updater-manager/std",
 	"nomad-home/std",
 	"nomad-da-bridge/std",
